Change function

+++ src/main/java/KCPM/Demo_Jenkins_Java/Function.java
@@ -10,7 +10,7 @@ public class Function {
 	{
 		return a - b;
 	}
-	public double NhanHaiSoNguyen(int a, int b)
+	public int NhanHaiSoNguyen(int a, int b)
 	{
 		return a * b;
 	}
